am62ax: VHWA: Changing pointer type for optimization
- This is done in order to avoid a format conversion from upper layer
  to lower layer, and back again when writing to configuration RAM.
- This optimization saves > 200us of CPU time, as well as intermediate
    buffer memory.

Fixes: ISS-504, ISS-510

+++ test/drivers/vhwa/viss/vhwa_viss_cfg.h
@@ -1280,7 +1280,7 @@ Fcp_comDecomLutConfig gCfaComDecomLut[] =
     },
 };
 
-uint32_t gNsf4HistLut [NSF4_HISTOGRAM_LUT_SIZE] =
+uint16_t gNsf4HistLut [NSF4_HISTOGRAM_LUT_SIZE] =
 {
     #include "nsf4_hist_lut.txt"
 };
@@ -2797,11 +2797,11 @@ H3a_Config gH3aCfg[] =
     },
 };
 
-int32_t gYeeLut[] = {
+int16_t gYeeLut[] = {
     #include "yee_lut.txt"
 };
 
-int32_t gYeeLutLca[] = {
+int16_t gYeeLutLca[] = {
     #include "yee_lut_lca.txt"
 };
